+++ Statsd.c
@@ -1,3 +1,3 @@
-int Statsd_getMetric(const char* metricName) {
-  return 15;
+double Statsd_getMetric(const char* metricName) {
+  return 15.0;
 }
